# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Blog API",
    "version": "1.0.0",
    "description": "REST API for the Blog backend (Express 5, MongoDB/Mongoose, JWT). All JSON routes are under `/api/v1`.",
    "license": {
      "name": "Apache-2.0"
    }
  },
  "servers": [
    {
      "url": "/"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT access token from login/register response."
      },
      "cookieRefreshToken": {
        "type": "apiKey",
        "in": "cookie",
        "name": "refreshToken",
        "description": "HTTP-only JWT refresh cookie set on login/register."
      }
    },
    "schemas": {
      "ValidationError": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "example": "ValidationError"
          },
          "errors": {
            "type": "object",
            "additionalProperties": true,
            "description": "express-validator mapped errors"
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/v1/blogs/user/{userId}": {
      "get": {
        "tags": [
          "Blogs"
        ],
        "summary": "List blogs by author user ID (admin only)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 50
            }
          },
          {
            "in": "query",
            "name": "offset",
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated blogs for user"
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Not admin"
          }
        }
      }
    },
...
    "/api/v1/users/current": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Get current authenticated user",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Current user profile"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      },
      "put": {
        "tags": [
          "Users"
        ],
        "summary": "Update current user profile",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": {
                    "type": "string",
                    "maxLength": 20
                  },
                  "email": {
                    "type": "string",
                    "format": "email",
                    "maxLength": 50
                  },
                  "password": {
                    "type": "string",
                    "minLength": 8
                  },
                  "first_name": {
                    "type": "string",
                    "maxLength": 20
                  },
                  "last_name": {
                    "type": "string",
                    "maxLength": 20
                  },
                  "website": {
                    "type": "string",
                    "format": "uri",
                    "maxLength": 100
                  },
                  "facebook": {
                    "type": "string",
                    "format": "uri",
                    "maxLength": 100
                  },
                  "instagram": {
                    "type": "string",
                    "format": "uri",
                    "maxLength": 100
                  },
                  "linkedin": {
                    "type": "string",
                    "format": "uri",
                    "maxLength": 100
                  },
                  "x": {
                    "type": "string",
                    "format": "uri",
                    "maxLength": 100
                  },
                  "youtube": {
                    "type": "string",
                    "format": "uri",
                    "maxLength": 100
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated"
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "delete": {
        "tags": [
          "Users"
        ],
        "summary": "Delete current user account",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "204": {
            "description": "Account deleted (no body)"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/api/v1/users": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "List users (admin only)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 50
            }
          },
          {
            "in": "query",
            "name": "offset",
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated users"
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Not admin"
          }
        }
      }
    },
    "/api/v1/users/{userId}": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Get user by ID (admin only)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Mongo ObjectId"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User document"
          },
          "400": {
            "description": "Invalid ID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Not admin"
          },
          "404": {
            "description": "User not found"
          }
        }
      },
      "delete": {
        "tags": [
          "Users"
        ],
        "summary": "Delete user by ID (admin only)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "User and related blogs removed (no body)"
          },
          "400": {
            "description": "Invalid ID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Not admin"
          },
          "404": {
            "description": "User not found (controller path)"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    }
  }
}